[Tizen] (TextAtlas) Let Text relative geometry use uint32_t indices
Due to the data format is uint16_t type, we cannot render well
if the length of text is bigger than 16384.
If we add underline or strokethrough, the number of text limitation
also be reduced.

To fix this issue, let we make indices what TextAtlas geometry use
uint32_t type, instead uint16_t type.

Since we support uint32_t* indices now a day, it will be okay.

+++ dali-toolkit/internal/text/rendering/atlas/atlas-mesh-factory.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Samsung Electronics Co., Ltd.
+ * Copyright (c) 2025 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -147,8 +147,8 @@ void AppendMesh(Toolkit::AtlasManager::Mesh2D&       first,
                         second.mIndices.Begin(),
                         second.mIndices.End());
 
-  for(Vector<unsigned short>::Iterator it    = first.mIndices.Begin() + indicesCount,
-                                       endIt = first.mIndices.End();
+  for(Vector<uint32_t>::Iterator it    = first.mIndices.Begin() + indicesCount,
+                                 endIt = first.mIndices.End();
       it != endIt;
       ++it)
   {
